+++ packages/subtree-updater/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- ignore fill batch failure in case where batch already filled
 - only prune tree after proof job is submitted to bullmq
 - update .env.example
 - add docker-compose file